Revert "Feature/auctdb creating sq lite db"
